+++ src/iter.h
@@ -1,8 +1,9 @@
+#include "base/zdefs.h"
+#include "zc/maps.h"
 #include <functional>
 #include <optional>
 #include <type_traits>
-#include "base/zdefs.h"
-#include "zc/maps.h"
+#include <stdint.h>
 
 // Iterates over every screen in the current region.
 // Callback function: void fn(mapscr* screen, int screen_index, unsigned int region_scr_x, unsigned int region_scr_x)
@@ -69,7 +70,7 @@ template<typename T, typename = std::enable_if_t<
 >>
 void for_every_ffc_in_region(T fn)
 {
-	for (int screen_index = 0; screen_index < 128; screen_index++)
+	for (uint8_t screen_index = 0; screen_index < 128; screen_index++)
 	{
 		if (is_in_current_region(screen_index))
 		{
@@ -77,9 +78,9 @@ void for_every_ffc_in_region(T fn)
 			int screen_index_offset = get_region_screen_index_offset(screen_index);
 
 			int c = screen->numFFC();
-			for (int i = 0; i < c; i++)
+			for (uint8_t i = 0; i < c; i++)
 			{
-				int id = screen_index_offset * MAXFFCS + i;
+				uint16_t id = screen_index_offset * MAXFFCS + i;
 				ffc_handle_t ffc_handle = {screen, screen_index, id, i, &screen->ffcs[i]};
 				fn(ffc_handle);
 			}
@@ -95,7 +96,7 @@ template<typename T, typename = std::enable_if_t<
 >>
 void for_some_ffcs_in_region(T fn)
 {
-	for (int screen_index = 0; screen_index < 128; screen_index++)
+	for (uint8_t screen_index = 0; screen_index < 128; screen_index++)
 	{
 		if (is_in_current_region(screen_index))
 		{
@@ -103,9 +104,9 @@ void for_some_ffcs_in_region(T fn)
 			int screen_index_offset = get_region_screen_index_offset(screen_index);
 
 			int c = screen->numFFC();
-			for (int i = 0; i < c; i++)
+			for (uint8_t i = 0; i < c; i++)
 			{
-				int id = screen_index_offset * MAXFFCS + i;
+				uint16_t id = screen_index_offset * MAXFFCS + i;
 				ffc_handle_t ffc_handle = {screen, screen_index, id, i, &screen->ffcs[i]};
 				if (!fn(ffc_handle)) return;
 			}
@@ -121,7 +122,7 @@ template<typename T, typename = std::enable_if_t<
 >>
 std::optional<ffc_handle_t> find_ffc_in_region(T fn)
 {
-	for (int screen_index = 0; screen_index < 128; screen_index++)
+	for (uint8_t screen_index = 0; screen_index < 128; screen_index++)
 	{
 		if (is_in_current_region(screen_index))
 		{
@@ -129,9 +130,9 @@ std::optional<ffc_handle_t> find_ffc_in_region(T fn)
 			int screen_index_offset = get_region_screen_index_offset(screen_index);
 
 			int c = screen->numFFC();
-			for (int i = 0; i < c; i++)
+			for (uint8_t i = 0; i < c; i++)
 			{
-				int id = screen_index_offset * MAXFFCS + i;
+				uint16_t id = screen_index_offset * MAXFFCS + i;
 				ffc_handle_t ffc_handle = {screen, screen_index, id, i, &screen->ffcs[i]};
 				if (fn(ffc_handle)) return ffc_handle;
 			}
